requirements.txt
apt-clone==0.2.1
apt-xapian-index==0.47
apturl==0.5.2
awscli==1.11.13
backcall==0.1.0
bleach==2.1.4
blinker==1.3
blis==0.2.4
boto==2.49.0
boto3==1.9.7
botocore==1.12.7
Brlapi==0.6.4
bz2file==0.98
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.7
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
cymem==2.0.2
decorator==4.3.0
defer==1.0.6
defusedxml==0.5.0
dirspec==13.10
docutils==0.14
entrypoints==0.2.3
es-core-news-sm==2.1.0
ez-setup==0.9
fasttext==0.8.3
future==0.17.1
gensim==3.5.0
html5lib==1.0.1
httplib2==0.9.1
idna==2.8
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.3
murmurhash==1.0.2
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.3
notebook==5.6.0
numpy==1.15.2
oauthlib==1.0.3
onboard==1.2.0
oneconf==0.3.9
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==3.1.2
piston-mini-client==0.7.5
plac==0.9.6
plotly==3.2.1
pluggy==0.9.0
preshed==2.0.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
psutil==3.4.2
ptyprocess==0.6.0
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyICU==1.9.2
PyJWT==1.3.0
pyparsing==2.2.1
pytho==0.0.1
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.7.3
python-debian==0.1.27
python-jsonrpc-server==0.1.2
python-language-server==0.26.1
python-xapp==1.0.0
pytz==2018.5
pyxdg==0.25
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
reportlab==3.3.0
requests==2.21.0
retrying==1.3.3
roman==2.0.0
rsa==3.2.3
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
sessioninstaller==0.0.0
setproctitle==1.1.8
simplegeneric==0.8.1
six==1.11.0
smart-open==1.7.1
spacy==2.1.0
srsly==0.0.5
stop-words==2018.7.23
system-service==0.3
terminado==0.8.1
testpath==0.3.1
thin-client-config-agent==0.8
thinc==7.0.3
thinc-gpu-ops==0.0.4
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
urllib3==1.23
virtkey==0.63.0
virtualenv==15.0.1
wasabi==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xkit==0.0.0
